Errors fixed

+++ src/main/java/mk/ukim/finki/wpproekt/model/Bolnica.java
@@ -18,14 +18,14 @@ public class Bolnica {
     private Long bolnica_id;
     private String naziv;
     private String grad;
-    private Long broj;
+    private Integer broj;
     private String ulica;
     private String smetka_bolnica;
 
     public Bolnica() {
     }
 
-    public Bolnica(String naziv, String grad, Long broj, String ulica, String smetka_bolnica) {
+    public Bolnica(String naziv, String grad, Integer broj, String ulica, String smetka_bolnica) {
         this.naziv = naziv;
         this.grad = grad;
         this.broj = broj;
